requirements.txt
pymodbus==3.8.4
pyserial==3.5
huawei-solar==2.4.4
pyyaml==6.0.2